+++ object.go
@@ -15,15 +15,15 @@ type Objects interface {
 	SumBy(...string) Objects // 聚合计算
 	Meta() (*Meta, error)    // 延时摘要信息
 	// operate
-	Count() (int, error)             // 数目
-	All(result interface{}) error    // 保存搜索结果至
-	One(result interface{}) error    // 取一条记录
-	Create(insert interface{}) error // 插入一条记录
-	Update(record interface{}) error // 更改(输入struct或map) *** struct为覆盖更新，map为局部更新
-	UpdateOne(obj interface{}) error // 只确保更改一条(输入struct或map)
-	Delete() error                   // 删除
-	DeleteOne() error                // 删除一条记录
-	Sum(...string) ([]int, error)    // 立刻聚合计算
+	Count() (int, error)              // 数目
+	All(result interface{}) error     // 保存搜索结果至
+	One(result interface{}) error     // 取一条记录
+	Create(insert interface{}) error  // 插入一条记录
+	Update(record interface{}) error  // 更改(输入struct或map) *** struct为覆盖更新，map为局部更新
+	UpdateOne(obj interface{}) error  // 只确保更改一条(输入struct或map)
+	Delete() error                    // 删除
+	DeleteOne() error                 // 删除一条记录
+	Sum(...string) ([]float64, error) // 立刻聚合计算
 	// 事务操作
 	TLockUpdate(t Trans) error                 // 行锁
 	TCount(t Trans) (int, error)               // 数目
